Adding thoughts on paper story
I wanted to have a place to store our thoughts on what the
story of our paper will be. By version controlling this
we can more easily see how our thoughts evolve over time/revert
to old versions of the story if new ideas don't work out.

Specifically, this commit adds a file `story_outline.md` with
a rough outline of the paper story/figure ideas, and a second
file `papers_to_read.md`, with a list of papers that I think
would be good models for us.