Add back target-vs-background DE test
In the original contrastiveVI codebase, we could use our differential
expression test either for "target vs. target" testing or "target
vs. background" testing, for which salient latent variables would be
zeroed out for background samples. The latter functionality appears
to have been lost when we upgraded to the newer version of scvi-tools,
and this MR adds back in the capability.